Move PathProvider to Rapicgen.Core.External namespace
