+++ src/Core/ApiClientCodeGen.Core/Installer/DependencyInstaller.cs
@@ -1,5 +1,6 @@
 using System;
 using System.ComponentModel;
+using Rapicgen.Core.External;
 using Rapicgen.Core.Generators;
 using Rapicgen.Core.Logging;
 using Rapicgen.Core.Options.General;
