+++ src/Core/ApiClientCodeGen.Tests.Common/Fixtures/OpenApi3/Yaml/SwaggerCodeGeneratorFixture.cs
@@ -5,6 +5,7 @@ using Rapicgen.Core.Generators.Swagger;
 using Rapicgen.Core.Installer;
 using Rapicgen.Core.Options.General;
 using Moq;
+using Rapicgen.Core.External;
 
 namespace ApiClientCodeGen.Tests.Common.Fixtures.OpenApi3.Yaml
 {
